@@ -1,107 +1,100 @@
-import React, {Component} from 'react'
-import Card, {CardActions, CardContent} from 'material-ui/Card'
-import Button from 'material-ui/Button'
-import TextField from 'material-ui/TextField'
-import Typography from 'material-ui/Typography'
-import Icon from 'material-ui/Icon'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
+import React, {useState} from 'react'
+import Card from '@material-ui/core/Card'
+import CardActions from '@material-ui/core/CardActions'
+import CardContent from '@material-ui/core/CardContent'
+import Button from '@material-ui/core/Button'
+import TextField from '@material-ui/core/TextField'
+import Typography from '@material-ui/core/Typography'
+import Icon from '@material-ui/core/Icon'
+import { makeStyles } from '@material-ui/core/styles'
 import auth from './../auth/auth-helper'
 import {Redirect} from 'react-router-dom'
 import {signin} from './api-auth.js'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   card: {
     maxWidth: 600,
     margin: 'auto',
     textAlign: 'center',
-    marginTop: theme.spacing.unit * 5,
-    paddingBottom: theme.spacing.unit * 2
+    marginTop: theme.spacing(5),
+    paddingBottom: theme.spacing(2)
   },
   error: {
     verticalAlign: 'middle'
   },
   title: {
-    marginTop: theme.spacing.unit * 2,
+    marginTop: theme.spacing(2),
     color: theme.palette.openTitle
   },
   textField: {
-    marginLeft: theme.spacing.unit,
-    marginRight: theme.spacing.unit,
+    marginLeft: theme.spacing(1),
+    marginRight: theme.spacing(1),
     width: 300
   },
   submit: {
     margin: 'auto',
-    marginBottom: theme.spacing.unit * 2
+    marginBottom: theme.spacing(2)
   }
-})
+}))
 
-class Signin extends Component {
-  state = {
+export default function Signin(props) {
+  const classes = useStyles();
+  const [values, setValues] = useState({
       email: '',
       password: '',
       error: '',
       redirectToReferrer: false
-  }
+  })
 
-  clickSubmit = () => {
+  const clickSubmit = () => {
     const user = {
-      email: this.state.email || undefined,
-      password: this.state.password || undefined
+      email: values.email || undefined,
+      password: values.password || undefined
     }
 
     signin(user).then((data) => {
       if (data.error) {
-        this.setState({error: data.error})
+        setValues({ ...values, error: data.error})
       } else {
         auth.authenticate(data, () => {
-          this.setState({redirectToReferrer: true})
+          setValues({ ...values, error: '',redirectToReferrer: true})
         })
       }
     })
   }
 
-  handleChange = name => event => {
-    this.setState({[name]: event.target.value})
+  const handleChange = name => event => {
+    setValues({ ...values, [name]: event.target.value })
   }
 
-  render() {
-    const {classes} = this.props
-    const {from} = this.props.location.state || {
+  const {from} = props.location.state || {
       from: {
         pathname: '/'
       }
-    }
-    const {redirectToReferrer} = this.state
+  }
+  const {redirectToReferrer} = values
     if (redirectToReferrer) {
       return (<Redirect to={from}/>)
-    }
+  }
 
-    return (
+  return (
       <Card className={classes.card}>
         <CardContent>
-          <Typography type="headline" component="h2" className={classes.title}>
+          <Typography variant="h5" className={classes.title}>
             Sign In
           </Typography>
-          <TextField id="email" type="email" label="Email" className={classes.textField} value={this.state.email} onChange={this.handleChange('email')} margin="normal"/><br/>
-          <TextField id="password" type="password" label="Password" className={classes.textField} value={this.state.password} onChange={this.handleChange('password')} margin="normal"/>
+          <TextField id="email" type="email" label="Email" className={classes.textField} value={values.email} onChange={handleChange('email')} margin="normal"/><br/>
+          <TextField id="password" type="password" label="Password" className={classes.textField} value={values.password} onChange={handleChange('password')} margin="normal"/>
           <br/> {
-            this.state.error && (<Typography component="p" color="error">
+            values.error && (<Typography component="p" color="error">
               <Icon color="error" className={classes.error}>error</Icon>
-              {this.state.error}
+              {values.error}
             </Typography>)
           }
         </CardContent>
         <CardActions>
-          <Button color="primary" variant="raised" onClick={this.clickSubmit} className={classes.submit}>Submit</Button>
+          <Button color="primary" variant="contained" onClick={clickSubmit} className={classes.submit}>Submit</Button>
         </CardActions>
       </Card>
     )
-  }
 }
-
-Signin.propTypes = {
-  classes: PropTypes.object.isRequired
-}
-
-export default withStyles(styles)(Signin)
