@@ -1,26 +1,29 @@
-import React, {Component} from 'react'
-import Card, { CardHeader, CardContent, CardActions } from 'material-ui/Card'
-import Button from 'material-ui/Button'
-import TextField from 'material-ui/TextField'
-import Typography from 'material-ui/Typography'
-import Avatar from 'material-ui/Avatar'
-import Icon from 'material-ui/Icon'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
-import {create} from './api-post.js'
+import React, {useState, useEffect} from 'react'
 import auth from './../auth/auth-helper'
-import IconButton from 'material-ui/IconButton'
-import PhotoCamera from 'material-ui-icons/PhotoCamera'
+import Card from '@material-ui/core/Card'
+import CardHeader from '@material-ui/core/CardHeader'
+import CardContent from '@material-ui/core/CardContent'
+import CardActions from '@material-ui/core/CardActions'
+import Button from '@material-ui/core/Button'
+import TextField from '@material-ui/core/TextField'
+import Typography from '@material-ui/core/Typography'
+import Avatar from '@material-ui/core/Avatar'
+import Icon from '@material-ui/core/Icon'
+import PropTypes from 'prop-types'
+import {makeStyles} from '@material-ui/core/styles'
+import {create} from './api-post.js'
+import IconButton from '@material-ui/core/IconButton'
+import PhotoCamera from '@material-ui/icons/PhotoCamera'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   root: {
     backgroundColor: '#efefef',
-    padding: `${theme.spacing.unit*3}px 0px 1px`
+    padding: `${theme.spacing(3)}px 0px 1px`
   },
   card: {
     maxWidth:600,
     margin: 'auto',
-    marginBottom: theme.spacing.unit*3,
+    marginBottom: theme.spacing(3),
     backgroundColor: 'rgba(65, 150, 136, 0.09)',
     boxShadow: 'none'
   },
@@ -41,61 +44,61 @@ const styles = theme => ({
     display: 'none',
   },
   textField: {
-    marginLeft: theme.spacing.unit*2,
-    marginRight: theme.spacing.unit*2,
+    marginLeft: theme.spacing(2),
+    marginRight: theme.spacing(2),
     width: '90%'
   },
   submit: {
-    margin: theme.spacing.unit * 2
+    margin: theme.spacing(2)
   },
   filename:{
     verticalAlign: 'super'
   }
-})
+}))
 
-class NewPost extends Component {
-  state = {
+export default function NewPost (props){
+  const classes = useStyles()
+  const [values, setValues] = useState({
     text: '',
     photo: '',
     error: '',
     user: {}
-  }
-
-  componentDidMount = () => {
-    this.postData = new FormData()
-    this.setState({user: auth.isAuthenticated().user})
-  }
-  clickPost = () => {
-    const jwt = auth.isAuthenticated()
+  })
+  const jwt = auth.isAuthenticated()
+  useEffect(() => {
+    setValues({...values, user: auth.isAuthenticated().user})
+  }, [])
+  const clickPost = () => {
+    let postData = new FormData()
+    postData.append('text', values.text)
+    postData.append('photo', values.photo)
     create({
       userId: jwt.user._id
     }, {
       t: jwt.token
-    }, this.postData).then((data) => {
+    }, postData).then((data) => {
       if (data.error) {
-        this.setState({error: data.error})
+        setValues({...values, error: data.error})
       } else {
-        this.setState({text:'', photo: ''})
-        this.props.addUpdate(data)
+        setValues({...values, text:'', photo: ''})
+        props.addUpdate(data)
       }
     })
   }
-  handleChange = name => event => {
+  const handleChange = name => event => {
     const value = name === 'photo'
       ? event.target.files[0]
       : event.target.value
-    this.postData.set(name, value)
-    this.setState({ [name]: value })
+    setValues({...values, [name]: value })
   }
-  render() {
-    const {classes} = this.props
+
     return (<div className={classes.root}>
       <Card className={classes.card}>
       <CardHeader
             avatar={
-              <Avatar src={'/api/users/photo/'+this.state.user._id}/>
+              <Avatar src={'/api/users/photo/'+values.user._id}/>
             }
-            title={this.state.user.name}
+            title={values.user.name}
             className={classes.cardHeader}
           />
       <CardContent className={classes.cardContent}>
@@ -103,34 +106,32 @@ class NewPost extends Component {
             placeholder="Share your thoughts ..."
             multiline
             rows="3"
-            value={this.state.text}
-            onChange={this.handleChange('text')}
+            value={values.text}
+            onChange={handleChange('text')}
             className={classes.textField}
             margin="normal"
         />
-        <input accept="image/*" onChange={this.handleChange('photo')} className={classes.input} id="icon-button-file" type="file" />
+        <input accept="image/*" onChange={handleChange('photo')} className={classes.input} id="icon-button-file" type="file" />
         <label htmlFor="icon-button-file">
           <IconButton color="secondary" className={classes.photoButton} component="span">
             <PhotoCamera />
           </IconButton>
-        </label> <span className={classes.filename}>{this.state.photo ? this.state.photo.name : ''}</span>
-        { this.state.error && (<Typography component="p" color="error">
+        </label> <span className={classes.filename}>{values.photo ? values.photo.name : ''}</span>
+        { values.error && (<Typography component="p" color="error">
             <Icon color="error" className={classes.error}>error</Icon>
-              {this.state.error}
+              {values.error}
             </Typography>)
         }
       </CardContent>
       <CardActions>
-        <Button color="primary" variant="raised" disabled={this.state.text === ''} onClick={this.clickPost} className={classes.submit}>POST</Button>
+        <Button color="primary" variant="contained" disabled={values.text === ''} onClick={clickPost} className={classes.submit}>POST</Button>
       </CardActions>
     </Card>
   </div>)
-  }
+
 }
 
 NewPost.propTypes = {
-  classes: PropTypes.object.isRequired,
   addUpdate: PropTypes.func.isRequired
 }
 
-export default withStyles(styles)(NewPost)
