@@ -1,78 +1,77 @@
-import React, {Component} from 'react'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
-import Card from 'material-ui/Card'
-import Typography from 'material-ui/Typography'
-import Divider from 'material-ui/Divider'
+import React, {useState, useEffect} from 'react'
+import {makeStyles} from '@material-ui/core/styles'
+import Card from '@material-ui/core/Card'
+import Typography from '@material-ui/core/Typography'
+import Divider from '@material-ui/core/Divider'
 import auth from './../auth/auth-helper'
 import PostList from './PostList'
 import {listNewsFeed} from './api-post.js'
 import NewPost from './NewPost'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   card: {
     margin: 'auto',
     paddingTop: 0,
-    paddingBottom: theme.spacing.unit*3
+    paddingBottom: theme.spacing(3)
   },
   title: {
-    padding:`${theme.spacing.unit * 3}px ${theme.spacing.unit * 2.5}px ${theme.spacing.unit * 2}px`,
+    padding:`${theme.spacing(3)}px ${theme.spacing(2.5)}px ${theme.spacing(2)}px`,
     color: theme.palette.openTitle,
     fontSize: '1em'
   },
   media: {
     minHeight: 330
   }
-})
-class Newsfeed extends Component {
-  state = {
-      posts: []
-  }
-  loadPosts = () => {
-    const jwt = auth.isAuthenticated()
+}))
+export default function Newsfeed () {
+  const classes = useStyles()
+  const [posts, setPosts] = useState([])
+  const jwt = auth.isAuthenticated()
+
+  useEffect(() => {
+    const abortController = new AbortController()
+    const signal = abortController.signal
+
     listNewsFeed({
       userId: jwt.user._id
     }, {
       t: jwt.token
-    }).then((data) => {
+    }, signal).then((data) => {
       if (data.error) {
         console.log(data.error)
       } else {
-        this.setState({posts: data})
+        setPosts(data)
       }
     })
-  }
-  componentDidMount = () => {
-    this.loadPosts()
-  }
-  addPost = (post) => {
-    const updatedPosts = this.state.posts
+    return function cleanup(){
+      abortController.abort()
+    }
+
+  }, [])
+
+  const addPost = (post) => {
+    const updatedPosts = [...posts]
     updatedPosts.unshift(post)
-    this.setState({posts: updatedPosts})
+    console.log(updatedPosts)
+    setPosts(updatedPosts)
   }
-  removePost = (post) => {
-    const updatedPosts = this.state.posts
+  const removePost = (post) => {
+    const updatedPosts = [...posts]
     const index = updatedPosts.indexOf(post)
     updatedPosts.splice(index, 1)
-    this.setState({posts: updatedPosts})
+    setPosts(updatedPosts)
   }
-  render() {
-    const {classes} = this.props
+
     return (
       <Card className={classes.card}>
         <Typography type="title" className={classes.title}>
           Newsfeed
         </Typography>
         <Divider/>
-        <NewPost addUpdate={this.addPost}/>
+        <NewPost addUpdate={addPost}/>
         <Divider/>
-        <PostList removeUpdate={this.removePost} posts={this.state.posts}/>
+        <PostList removeUpdate={removePost} posts={posts}/>
       </Card>
     )
-  }
-}
-Newsfeed.propTypes = {
-  classes: PropTypes.object.isRequired
 }
 
-export default withStyles(styles)(Newsfeed)
