@@ -1,48 +1,50 @@
-import React, {Component} from 'react'
+import React, {useState} from 'react'
 import PropTypes from 'prop-types'
-import IconButton from 'material-ui/IconButton'
-import Button from 'material-ui/Button'
-import DeleteIcon from 'material-ui-icons/Delete'
-import Dialog, {DialogActions, DialogContent, DialogContentText, DialogTitle} from 'material-ui/Dialog'
+import IconButton from '@material-ui/core/IconButton'
+import Button from '@material-ui/core/Button'
+import DeleteIcon from '@material-ui/icons/Delete'
+import Dialog from '@material-ui/core/Dialog'
+import DialogActions from '@material-ui/core/DialogActions'
+import DialogContent from '@material-ui/core/DialogContent'
+import DialogContentText from '@material-ui/core/DialogContentText'
+import DialogTitle from '@material-ui/core/DialogTitle'
 import auth from './../auth/auth-helper'
 import {remove} from './api-user.js'
 import {Redirect, Link} from 'react-router-dom'
 
-class DeleteUser extends Component {
-  state = {
-    redirect: false,
-    open: false
+export default function DeleteUser(props) {
+  const [open, setOpen] = useState(false)
+  const [redirect, setRedirect] = useState(false)
+  
+  const jwt = auth.isAuthenticated()
+  const clickButton = () => {
+    setOpen(true)
   }
-  clickButton = () => {
-    this.setState({open: true})
-  }
-  deleteAccount = () => {
-    const jwt = auth.isAuthenticated()
+  const deleteAccount = () => { 
     remove({
-      userId: this.props.userId
+      userId: props.userId
     }, {t: jwt.token}).then((data) => {
-      if (data.error) {
+      if (data && data.error) {
         console.log(data.error)
       } else {
         auth.signout(() => console.log('deleted'))
-        this.setState({redirect: true})
+        setRedirect(true)
       }
     })
   }
-  handleRequestClose = () => {
-    this.setState({open: false})
+  const handleRequestClose = () => {
+    setOpen(false)
   }
-  render() {
-    const redirect = this.state.redirect
+
     if (redirect) {
       return <Redirect to='/'/>
     }
     return (<span>
-      <IconButton aria-label="Delete" onClick={this.clickButton} color="secondary">
+      <IconButton aria-label="Delete" onClick={clickButton} color="secondary">
         <DeleteIcon/>
       </IconButton>
 
-      <Dialog open={this.state.open} onClose={this.handleRequestClose}>
+      <Dialog open={open} onClose={handleRequestClose}>
         <DialogTitle>{"Delete Account"}</DialogTitle>
         <DialogContent>
           <DialogContentText>
@@ -50,18 +52,16 @@ class DeleteUser extends Component {
           </DialogContentText>
         </DialogContent>
         <DialogActions>
-          <Button onClick={this.handleRequestClose} color="primary">
+          <Button onClick={handleRequestClose} color="primary">
             Cancel
           </Button>
-          <Button onClick={this.deleteAccount} color="secondary" autoFocus="autoFocus">
+          <Button onClick={deleteAccount} color="secondary" autoFocus="autoFocus">
             Confirm
           </Button>
         </DialogActions>
       </Dialog>
     </span>)
-  }
 }
 DeleteUser.propTypes = {
   userId: PropTypes.string.isRequired
 }
-export default DeleteUser
