@@ -1,33 +1,36 @@
-import React, {Component} from 'react'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
-import Paper from 'material-ui/Paper'
-import List, {ListItem, ListItemAvatar, ListItemSecondaryAction, ListItemText} from 'material-ui/List'
-import Avatar from 'material-ui/Avatar'
-import Button from 'material-ui/Button'
-import IconButton from 'material-ui/IconButton'
-import Typography from 'material-ui/Typography'
+import React, {useState, useEffect} from 'react'
+import {makeStyles} from '@material-ui/core/styles'
+import Paper from '@material-ui/core/Paper'
+import List from '@material-ui/core/List'
+import ListItem from '@material-ui/core/ListItem'
+import ListItemAvatar from '@material-ui/core/ListItemAvatar'
+import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction' 
+import ListItemText from '@material-ui/core/ListItemText' 
+import Avatar from '@material-ui/core/Avatar'
+import Button from '@material-ui/core/Button'
+import IconButton from '@material-ui/core/IconButton'
+import Typography from '@material-ui/core/Typography'
 import {Link} from 'react-router-dom'
 import {findPeople, follow} from './api-user.js'
 import auth from './../auth/auth-helper'
-import Snackbar from 'material-ui/Snackbar'
-import ViewIcon from 'material-ui-icons/Visibility'
+import Snackbar from '@material-ui/core/Snackbar'
+import ViewIcon from '@material-ui/icons/Visibility'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   root: theme.mixins.gutters({
-    padding: theme.spacing.unit,
+    padding: theme.spacing(1),
     margin: 0
   }),
   title: {
-    margin: `${theme.spacing.unit * 3}px ${theme.spacing.unit}px ${theme.spacing.unit * 2}px`,
+    margin: `${theme.spacing(3)}px ${theme.spacing(1)}px ${theme.spacing(2)}px`,
     color: theme.palette.openTitle,
     fontSize: '1em'
   },
   avatar: {
-    marginRight: theme.spacing.unit * 1
+    marginRight: theme.spacing(1)
   },
   follow: {
-    right: theme.spacing.unit * 2
+    right: theme.spacing(2)
   },
   snack: {
     color: theme.palette.protectedTitle
@@ -35,54 +38,62 @@ const styles = theme => ({
   viewButton: {
     verticalAlign: 'middle'
   }
-})
-class FindPeople extends Component {
-  state = {
-      users: [],
-      open: false
-  }
-  componentDidMount = () => {
-    const jwt = auth.isAuthenticated()
+}))
+
+export default function FindPeople() {
+  const classes = useStyles()
+  const [values, setValues] = useState({
+    users: [],
+    open: false,
+    followMessage: ''
+  })
+  const jwt = auth.isAuthenticated()
+
+  useEffect(() => {
+    const abortController = new AbortController()
+    const signal = abortController.signal
+
     findPeople({
       userId: jwt.user._id
     }, {
       t: jwt.token
-    }).then((data) => {
-      if (data.error) {
+    }, signal).then((data) => {
+      if (data && data.error) {
         console.log(data.error)
       } else {
-        this.setState({users: data})
+        setValues({...values, users:data})
       }
     })
-  }
-  clickFollow = (user, index) => {
-    const jwt = auth.isAuthenticated()
+    return function cleanup(){
+      abortController.abort()
+    }
+
+  }, [])
+  const clickFollow = (user, index) => {
     follow({
       userId: jwt.user._id
     }, {
       t: jwt.token
     }, user._id).then((data) => {
       if (data.error) {
-        this.setState({error: data.error})
+        console.log(data.error)
       } else {
-        let toFollow = this.state.users
+        let toFollow = values.users
         toFollow.splice(index, 1)
-        this.setState({users: toFollow, open: true, followMessage: `Following ${user.name}!`})
+        setValues({...values, users: toFollow, open: true, followMessage: `Following ${user.name}!`})
       }
     })
   }
-  handleRequestClose = (event, reason) => {
-    this.setState({ open: false })
+  const handleRequestClose = (event, reason) => {
+    setValues({...values, open: false })
   }
-  render() {
-    const {classes} = this.props
     return (<div>
       <Paper className={classes.root} elevation={4}>
         <Typography type="title" className={classes.title}>
           Who to follow
         </Typography>
         <List>
-          {this.state.users.map((item, i) => {
+          {values.users.map((item, i) => {
               return <span key={i}>
                 <ListItem>
                   <ListItemAvatar className={classes.avatar}>
@@ -91,11 +102,11 @@ class FindPeople extends Component {
                   <ListItemText primary={item.name}/>
                   <ListItemSecondaryAction className={classes.follow}>
                     <Link to={"/user/" + item._id}>
-                      <IconButton variant="raised" color="secondary" className={classes.viewButton}>
+                      <IconButton variant="contained" color="secondary" className={classes.viewButton}>
                         <ViewIcon/>
                       </IconButton>
                     </Link>
-                    <Button aria-label="Follow" variant="raised" color="primary" onClick={this.clickFollow.bind(this, item, i)}>
+                    <Button aria-label="Follow" variant="contained" color="primary" onClick={()=> {clickFollow(item, i)}}>
                       Follow
                     </Button>
                   </ListItemSecondaryAction>
@@ -110,17 +121,10 @@ class FindPeople extends Component {
             vertical: 'bottom',
             horizontal: 'right',
           }}
-          open={this.state.open}
-          onClose={this.handleRequestClose}
+          open={values.open}
+          onClose={handleRequestClose}
           autoHideDuration={6000}
-          message={<span className={classes.snack}>{this.state.followMessage}</span>}
+          message={<span className={classes.snack}>{values.followMessage}</span>}
       />
     </div>)
-  }
 }
-
-FindPeople.propTypes = {
-  classes: PropTypes.object.isRequired
-}
-
-export default withStyles(styles)(FindPeople)
