@@ -1,48 +1,39 @@
-import React, {Component} from 'react'
+import React, {useState} from 'react'
 import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
-import Button from 'material-ui/Button'
-import {Link} from 'react-router-dom'
-import AppBar from 'material-ui/AppBar'
-import Typography from 'material-ui/Typography'
-import Tabs, { Tab } from 'material-ui/Tabs'
+import AppBar from '@material-ui/core/AppBar'
+import Typography from '@material-ui/core/Typography'
+import Tabs from '@material-ui/core/Tabs'
+import Tab from '@material-ui/core/Tab'
 import FollowGrid from './../user/FollowGrid'
 import PostList from './../post/PostList'
 
-class ProfileTabs extends Component {
-  state = {
-    tab: 0,
-    posts: []
+export default function ProfileTabs ( props ){
+  const [tab, setTab] = useState(0)
+
+  const handleTabChange = (event, value) => {
+    setTab(value)
   }
 
-  componentWillReceiveProps = (props) => {
-    this.setState({tab:0})
-  }
-  handleTabChange = (event, value) => {
-    this.setState({ tab: value })
-  }
-
-  render() {
     return (
     <div>
         <AppBar position="static" color="default">
           <Tabs
-            value={this.state.tab}
-            onChange={this.handleTabChange}
+            value={tab}
+            onChange={handleTabChange}
             indicatorColor="primary"
             textColor="primary"
-            fullWidth
+            variant="fullWidth"
           >
             <Tab label="Posts" />
             <Tab label="Following" />
             <Tab label="Followers" />
           </Tabs>
         </AppBar>
-       {this.state.tab === 0 && <TabContainer><PostList removeUpdate={this.props.removePostUpdate} posts={this.props.posts}/></TabContainer>}
-       {this.state.tab === 1 && <TabContainer><FollowGrid people={this.props.user.following}/></TabContainer>}
-       {this.state.tab === 2 && <TabContainer><FollowGrid people={this.props.user.followers}/></TabContainer>}
+       {tab === 0 && <TabContainer><PostList removeUpdate={props.removePostUpdate} posts={props.posts}/></TabContainer>}
+       {tab === 1 && <TabContainer><FollowGrid people={props.user.following}/></TabContainer>}
+       {tab === 2 && <TabContainer><FollowGrid people={props.user.followers}/></TabContainer>}
     </div>)
-  }
+  
 }
 
 ProfileTabs.propTypes = {
@@ -62,5 +53,3 @@ const TabContainer = (props) => {
 TabContainer.propTypes = {
   children: PropTypes.node.isRequired
 }
-
-export default ProfileTabs
