@@ -1,51 +1,59 @@
-import React, {Component} from 'react'
-import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
-import Paper from 'material-ui/Paper'
-import List, {ListItem, ListItemAvatar, ListItemSecondaryAction, ListItemText} from 'material-ui/List'
-import Avatar from 'material-ui/Avatar'
-import IconButton from 'material-ui/IconButton'
-import Typography from 'material-ui/Typography'
-import ArrowForward from 'material-ui-icons/ArrowForward'
-import Person from 'material-ui-icons/Person'
+import React, {useState, useEffect} from 'react'
+import { makeStyles } from '@material-ui/core/styles'
+import Paper from '@material-ui/core/Paper'
+import List from '@material-ui/core/List'
+import ListItem from '@material-ui/core/ListItem'
+import ListItemAvatar from '@material-ui/core/ListItemAvatar'
+import ListItemSecondaryAction from '@material-ui/core/ListItemSecondaryAction'
+import ListItemText from '@material-ui/core/ListItemText'
+import Avatar from '@material-ui/core/Avatar'
+import IconButton from '@material-ui/core/IconButton'
+import Typography from '@material-ui/core/Typography'
+import ArrowForward from '@material-ui/icons/ArrowForward'
+import Person from '@material-ui/icons/Person'
 import {Link} from 'react-router-dom'
 import {list} from './api-user.js'
 
-const styles = theme => ({
+const useStyles = makeStyles(theme => ({
   root: theme.mixins.gutters({
-    padding: theme.spacing.unit,
-    margin: theme.spacing.unit * 5
+    padding: theme.spacing(1),
+    margin: theme.spacing(5)
   }),
   title: {
-    margin: `${theme.spacing.unit * 4}px 0 ${theme.spacing.unit * 2}px`,
+    margin: `${theme.spacing(4)}px 0 ${theme.spacing(2)}px`,
     color: theme.palette.openTitle
   }
-})
+}))
 
-class Users extends Component {
-  state = {
-      users: []
-  }
+export default function Users() { 
+  const classes = useStyles();
+  const [users, setUsers] = useState([])
 
-  componentDidMount() {
-    list().then((data) => {
-      if (data.error) {
+  useEffect(() => {
+    const abortController = new AbortController()
+    const signal = abortController.signal
+
+    list(signal).then((data) => {
+      if (data && data.error) {
         console.log(data.error)
       } else {
-        this.setState({users: data})
+        setUsers(data)
       }
     })
-  }
 
-  render() {
-    const {classes} = this.props
+    return function cleanup(){
+      abortController.abort()
+    }
+  }, [])
+
+
     return (
       <Paper className={classes.root} elevation={4}>
-        <Typography type="title" className={classes.title}>
+        <Typography variant="h6" className={classes.title}>
           All Users
         </Typography>
         <List dense>
-         {this.state.users.map((item, i) => {
+         {users.map((item, i) => {
           return <Link to={"/user/" + item._id} key={i}>
                     <ListItem button>
                       <ListItemAvatar>
@@ -66,11 +74,4 @@ class Users extends Component {
         </List>
       </Paper>
     )
-  }
 }
-
-Users.propTypes = {
-  classes: PropTypes.object.isRequired
-}
-
-export default withStyles(styles)(Users)
