@@ -1,22 +1,18 @@
-import React, {Component} from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
 import Post from './Post'
 
-class PostList extends Component {
-  render() {
+export default function PostList (props) {
     return (
       <div style={{marginTop: '24px'}}>
-        {this.props.posts.map((item, i) => {
-            return <Post post={item} key={i} onRemove={this.props.removeUpdate}/>
+        {props.posts.map((item, i) => {
+            return <Post post={item} key={i} onRemove={props.removeUpdate}/>
           })
         }
       </div>
     )
-  }
 }
 PostList.propTypes = {
   posts: PropTypes.array.isRequired,
   removeUpdate: PropTypes.func.isRequired
 }
-export default PostList
