@@ -1,27 +1,23 @@
-import React, {Component} from 'react'
+import React from 'react'
 import PropTypes from 'prop-types'
-import {withStyles} from 'material-ui/styles'
-import Button from 'material-ui/Button'
+import Button from '@material-ui/core/Button'
 import {unfollow, follow} from './api-user.js'
 
-class FollowProfileButton extends Component {
-  followClick = () => {
-    this.props.onButtonClick(follow)
+export default function FollowProfileButton (props) {
+  const followClick = () => {
+    props.onButtonClick(follow)
   }
-  unfollowClick = () => {
-    this.props.onButtonClick(unfollow)
+  const unfollowClick = () => {
+    props.onButtonClick(unfollow)
   }
-  render() {
     return (<div>
-      { this.props.following
-        ? (<Button variant="raised" color="secondary" onClick={this.unfollowClick}>Unfollow</Button>)
-        : (<Button variant="raised" color="primary" onClick={this.followClick}>Follow</Button>)
+      { props.following
+        ? (<Button variant="contained" color="secondary" onClick={unfollowClick}>Unfollow</Button>)
+        : (<Button variant="contained" color="primary" onClick={followClick}>Follow</Button>)
       }
     </div>)
-  }
 }
 FollowProfileButton.propTypes = {
   following: PropTypes.bool.isRequired,
   onButtonClick: PropTypes.func.isRequired
 }
-export default FollowProfileButton
