@@ -1,5 +1,6 @@
-const create = (user) => {
-  return fetch('/api/users/', {
+const create = async (user) => {
+  try {
+    let response = await fetch('/api/users/', {
       method: 'POST',
       headers: {
         'Accept': 'application/json',
@@ -7,103 +8,121 @@ const create = (user) => {
       },
       body: JSON.stringify(user)
     })
-    .then((response) => {
-      return response.json()
-    }).catch((err) => console.log(err))
-}
-
-const list = () => {
-  return fetch('/api/users/', {
-    method: 'GET',
-  }).then(response => {
-    return response.json()
-  }).catch((err) => console.log(err))
-}
-
-const read = (params, credentials) => {
-  return fetch('/api/users/' + params.userId, {
-    method: 'GET',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    }
-  }).then((response) => {
-    return response.json()
-  }).catch((err) => console.log(err))
-}
-
-const update = (params, credentials, user) => {
-  return fetch('/api/users/' + params.userId, {
-    method: 'PUT',
-    headers: {
-      'Accept': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    },
-    body: user
-  }).then((response) => {
-    return response.json()
-  }).catch((err) => {
+    return await response.json()
+  } catch(err) {
     console.log(err)
-  })
+  }
 }
 
-const remove = (params, credentials) => {
-  return fetch('/api/users/' + params.userId, {
-    method: 'DELETE',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    }
-  }).then((response) => {
-    return response.json()
-  }).catch((err) => console.log(err))
-}
-
-const follow = (params, credentials, followId) => {
-  return fetch('/api/users/follow/', {
-    method: 'PUT',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    },
-    body: JSON.stringify({userId:params.userId, followId: followId})
-  }).then((response) => {
-    return response.json()
-  }).catch((err) => {
+const list = async (signal) => {
+  try {
+    let response = await fetch('/api/users/', {
+      method: 'GET',
+      signal: signal
+    })
+    return await response.json()
+  } catch(err) {
     console.log(err)
-  })
+  }
 }
 
-const unfollow = (params, credentials, unfollowId) => {
-  return fetch('/api/users/unfollow/', {
-    method: 'PUT',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    },
-    body: JSON.stringify({userId:params.userId, unfollowId: unfollowId})
-  }).then((response) => {
-    return response.json()
-  }).catch((err) => {
+const read = async (params, credentials, signal) => {
+  try {
+    let response = await fetch('/api/users/' + params.userId, {
+      method: 'GET',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      }
+  })
+    return await response.json()
+  } catch(err) {
     console.log(err)
-  })
+  }
 }
 
-const findPeople = (params, credentials) => {
-  return fetch('/api/users/findpeople/' + params.userId, {
-    method: 'GET',
-    headers: {
-      'Accept': 'application/json',
-      'Content-Type': 'application/json',
-      'Authorization': 'Bearer ' + credentials.t
-    }
-  }).then((response) => {
-    return response.json()
-  }).catch((err) => console.log(err))
+const update = async (params, credentials, user) => {
+  try {
+    let response = await fetch('/api/users/' + params.userId, {
+      method: 'PUT',
+      headers: {
+        'Accept': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      },
+      body: user
+    })
+    return await response.json()
+  } catch(err) {
+    console.log(err)
+  }
+}
+
+const remove = async (params, credentials) => {
+  try {
+    let response = await fetch('/api/users/' + params.userId, {
+      method: 'DELETE',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      }
+    })
+    return await response.json()
+  } catch(err) {
+    console.log(err)
+  }
+}
+
+const follow = async (params, credentials, followId) => {
+  try {
+    let response = await fetch('/api/users/follow/', {
+      method: 'PUT',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      },
+      body: JSON.stringify({userId:params.userId, followId: followId})
+    })
+    return await response.json()
+  } catch(err) {
+    console.log(err)
+  }
+}
+
+const unfollow = async (params, credentials, unfollowId) => {
+  try {
+    let response = await fetch('/api/users/unfollow/', {
+      method: 'PUT',
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      },
+      body: JSON.stringify({userId:params.userId, unfollowId: unfollowId})
+    })
+    return await response.json()
+  } catch(err) {
+    console.log(err)
+  }
+}
+
+const findPeople = async (params, credentials, signal) => {
+  try {
+    let response = await fetch('/api/users/findpeople/' + params.userId, {
+      method: 'GET',
+      signal: signal,
+      headers: {
+        'Accept': 'application/json',
+        'Content-Type': 'application/json',
+        'Authorization': 'Bearer ' + credentials.t
+      }
+    })    
+    return await response.json()
+  } catch(err) {
+    console.log(err)
+  }
 }
 
 export {
