@@ -8,6 +8,7 @@ const create = (req, res, next) => {
   let form = new formidable.IncomingForm()
   form.keepExtensions = true
   form.parse(req, (err, fields, files) => {
+    console.log(err)
     if (err) {
       return res.status(400).json({
         error: "Image could not be uploaded"
@@ -30,61 +31,64 @@ const create = (req, res, next) => {
   })
 }
 
-const postByID = (req, res, next, id) => {
-  Post.findById(id).populate('postedBy', '_id name').exec((err, post) => {
-    if (err || !post)
+const postByID = async (req, res, next, id) => {
+  try{
+    let post = await Post.findById(id).populate('postedBy', '_id name').exec()
+    if (!post)
       return res.status('400').json({
         error: "Post not found"
       })
     req.post = post
     next()
-  })
+  }catch(err){
+    return res.status('400').json({
+      error: "Could not retrieve use post"
+    })
+  }
 }
 
-const listByUser = (req, res) => {
-  Post.find({postedBy: req.profile._id})
-  .populate('comments', 'text created')
-  .populate('comments.postedBy', '_id name')
-  .populate('postedBy', '_id name')
-  .sort('-created')
-  .exec((err, posts) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+const listByUser = async (req, res) => {
+  try{
+    let posts = await Post.find({postedBy: req.profile._id})
+                          .populate('comments.postedBy', '_id name')
+                          .populate('postedBy', '_id name')
+                          .sort('-created')
+                          .exec()
     res.json(posts)
-  })
+  }catch(err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
-const listNewsFeed = (req, res) => {
+const listNewsFeed = async (req, res) => {
   let following = req.profile.following
   following.push(req.profile._id)
-  Post.find({postedBy: { $in : req.profile.following } })
-  .populate('comments', 'text created')
-  .populate('comments.postedBy', '_id name')
-  .populate('postedBy', '_id name')
-  .sort('-created')
-  .exec((err, posts) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+  try{
+    let posts = await Post.find({postedBy: { $in : req.profile.following } })
+                          .populate('comments.postedBy', '_id name')
+                          .populate('postedBy', '_id name')
+                          .sort('-created')
+                          .exec()
     res.json(posts)
-  })
+  }catch(err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
-const remove = (req, res) => {
+const remove = async (req, res) => {
   let post = req.post
-    post.remove((err, deletedPost) => {
-      if (err) {
-        return res.status(400).json({
-          error: errorHandler.getErrorMessage(err)
-        })
-      }
-      res.json(deletedPost)
+  try{
+    let deletedPost = post.remove()
+    res.json(deletedPost)
+  }catch(err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
     })
+  }
 }
 
 const photo = (req, res, next) => {
@@ -92,59 +96,56 @@ const photo = (req, res, next) => {
     return res.send(req.post.photo.data)
 }
 
-const like = (req, res) => {
-  Post.findByIdAndUpdate(req.body.postId, {$push: {likes: req.body.userId}}, {new: true})
-  .exec((err, result) => {
-    if (err) {
+const like = async (req, res) => {
+  try{
+    let result = await Post.findByIdAndUpdate(req.body.postId, {$push: {likes: req.body.userId}}, {new: true})
+    res.json(result)
+  }catch(err){
       return res.status(400).json({
         error: errorHandler.getErrorMessage(err)
       })
-    }
-    res.json(result)
-  })
+  }
 }
 
-const unlike = (req, res) => {
-  Post.findByIdAndUpdate(req.body.postId, {$pull: {likes: req.body.userId}}, {new: true})
-  .exec((err, result) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+const unlike = async (req, res) => {
+  try{
+    let result = await Post.findByIdAndUpdate(req.body.postId, {$pull: {likes: req.body.userId}}, {new: true})
     res.json(result)
-  })
+  }catch(err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
-
-const comment = (req, res) => {
+const comment = async (req, res) => {
   let comment = req.body.comment
   comment.postedBy = req.body.userId
-  Post.findByIdAndUpdate(req.body.postId, {$push: {comments: comment}}, {new: true})
-  .populate('comments.postedBy', '_id name')
-  .populate('postedBy', '_id name')
-  .exec((err, result) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+  try{
+    let result = await Post.findByIdAndUpdate(req.body.postId, {$push: {comments: comment}}, {new: true})
+                            .populate('comments.postedBy', '_id name')
+                            .populate('postedBy', '_id name')
+                            .exec()
     res.json(result)
-  })
+  }catch(err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
-const uncomment = (req, res) => {
+const uncomment = async (req, res) => {
   let comment = req.body.comment
-  Post.findByIdAndUpdate(req.body.postId, {$pull: {comments: {_id: comment._id}}}, {new: true})
-  .populate('comments.postedBy', '_id name')
-  .populate('postedBy', '_id name')
-  .exec((err, result) => {
-    if (err) {
-      return res.status(400).json({
-        error: errorHandler.getErrorMessage(err)
-      })
-    }
+  try{
+    let result = await Post.findByIdAndUpdate(req.body.postId, {$pull: {comments: {_id: comment._id}}}, {new: true})
+                          .populate('comments.postedBy', '_id name')
+                          .populate('postedBy', '_id name')
+                          .exec()
     res.json(result)
-  })
+  }catch(err){
+    return res.status(400).json({
+      error: errorHandler.getErrorMessage(err)
+    })
+  }
 }
 
 const isPoster = (req, res, next) => {
