@@ -7,7 +7,6 @@ const config = {
     mode: "development",
     devtool: 'eval-source-map',
     entry: [
-        'react-hot-loader/patch',
         'webpack-hot-middleware/client?reload=true',
         path.join(CURRENT_WORKING_DIR, 'client/main.js')
     ],
@@ -30,10 +29,16 @@ const config = {
                 use: 'file-loader'
             }
         ]
-    },  plugins: [
+    },  
+    plugins: [
           new webpack.HotModuleReplacementPlugin(),
           new webpack.NoEmitOnErrorsPlugin()
-      ]
+    ],
+    resolve: {
+        alias: {
+          'react-dom': '@hot-loader/react-dom'
+        }
+    }
 }
 
 module.exports = config
